change mp_set_int/mp_init_set_int type, add mp_init_set_long etc for completeness

+++ bn_mp_init_set_int.c
@@ -12,8 +12,8 @@
  * SPDX-License-Identifier: Unlicense
  */
 
-/* initialize and set a digit */
-int mp_init_set_int(mp_int *a, unsigned long b)
+/* initialize and set int value */
+int mp_init_set_int(mp_int *a, unsigned int b)
 {
    int err;
    if ((err = mp_init(a)) != MP_OKAY) {
